2024 Add project utilities
Move `.gitignore` into this year, because it is likely to only apply to
this year; I am not yet sure how I will solve the remaining years, and
that's not something I need to worry about at the moment.

Define `ColumnLimit=100` as a compromise between WebKit's style having
no limit, and the "standard" 80 character limit; having a limit at all
helps clang-format be consistent, since there are fewer ways to organise
code.