2024 01 Parse Puzzle Input
It would be great to have `parseInput(std::string_view)`, since the
parameter is not modified, but unfortunately `std::istringstream` does
not work with the former, and I do not know a better way to parse the
input data.

`leftValue`, and `rightValue`'s default initialisation is overwritten
immediately when parsing the input string, but these elements are cheap
enough to copy where it is not worth to go against the norm, and not
initialise these variables.

Reserve space for for the individual lists to avoid multiple allocations
as they're filled; reserve in a local scope to limit the lifetime of
`listLength`: it's only used twice, immediately at creation.