2024 01 Add solver for part one
`solve(std::string)` creates a new copy of the input data; this is
intentional, since it has no use outside of this context, and that input
data should be passed as a forwarding reference, either with a temporary
r-value (as it is used in the unit tests), or by using `std::move()` to
"move" a named object onto the solver function.

`leftList`, and `rightList` are not `const` to allow them to be moved
onto `calculateDistances()`; `std::move(const&&)` actually calls
`std::move(const&)` [1]

[1] S. Meyers, "Item 23: Understand std::move and std::forward" in
Effective C++, Sebastopool, CA: O'Reilly Media, 2015, sec. 5.