2024 01 Refactor solver to include part two
`std::pair<std::size_t, std::size_t>` is well suited to be the output of
the `solve()`: we can refer to the individual results with
`result.first`, and `result.second`; do not create a new alias, since
that is an additional abstraction that does not add value.

`leftList`, and `rightList` can no longer be moved onto
`caluclateDistances()`, since they are needed also for `scoreLists()`;
Keep `std::move()` of boths lists in a previous commit as that was the
best move at the time, and every commit should be self-sufficient.