2024 02 Validate reports
`adjacentDifferences()` takes a `std::span` instead of a Report, because
it can, in theory, be used with any span, not just a `Report`; other
functions benefit from the additional context of having a `Report` as
argument.

Erase the first element of `adjacentDifferences` since it is simply the
first element of input, and not an actual difference to anything.

`isReportStable()` requires knowledge of the puzzle to understand its
meaning; this is acceptable, given this source code should be read with
that context anway.