2024 04 Count Xmas
Instead of checking only the beginning of the word "XMAS", testing from
both ends, that is "XMAS", or "SAMX" allows the grid to be tested only
for characters ahead, and avoid testing characters we already passed by;
this does have the issue that we must test the whole word, "XMAS" - if
we paid attention to only one end of the word, the beginning, we could
skip including that first character in the test string; spoiler warning,
this would be helpful for the second part of the puzzle; overall, I do
still believe the chosen tradeoff is worthwhile.

Each `isXmasInDirection()` first validates that there are enough
characters in that direction to make the word "XMAS"; those methods then
aggregate the characters to make a word, which is then checked to be
either "XMAS", or its reverse "SAMX".

Unfortunately, `std::string_view` must be a view into a contiguous
sequence of characters; this is why most checks must first allocate the
characters onto an array (we know the desired word length, so we can
define the length at compile time), to then read as a view.