2024 05 Parse page updates
The actual input always has a first section where the ordering rules are
defined; the two sections are separated by an empty line, which means
the ordering rules end with a double new-line character, i.e. "\n\n",
which must be ignored when parsing the data from the input.

Unit test also hypothetical cases, without ordering rules, to ensure
that those split characters are read, regardless of where they are found
in the input string.