2024 06 Parse guard's initial state
Coordinates increase rightwards, for `x`, and downwards, for `y`;
`Map` begins then at its top-leftmost corner, with `{.x=0, .y=0}`.

The input always includes one single guard; we will always find one of
the characters (`^>v<`), which will gives us both a position, and
orientation.

Both my example and puzzle inputs start with the guard facing up; I am
unaware if this is just a coincidence; if we could accept that the guard
always starts facing the same direction, we would initalize the guard
with that orientation, and skip `parseGuardOrientation()`.

The input includes `\n` characters that are not part of the map; rather
than extract these and then iterate over the cleaned data, account for
those extra characters, when converting the `index`.

Yes, iterating through the `string_view` multiple times, will result in
a slower solver; however, the impact on performance is not yet clear,
and keeping the parsers separate allows for a clearer code.