2024 06 Solve first puzzle
I don't like this solution, it is slow and not good; however, I spent
a lot of time misunderstanding the problem: I missed the part I had to
count ONLY distinct positions the guard visits; so I wanted a quick, and
dirty solution.

I will revisit this, once I have a solution for part two as well.