2024 07 Validate an equation
Rather than trying to brute force every possible combination of
operations, we can work backwards, starting from the end result, and
attempt to arrive at the first operand.

Prioritise checking if it is valid to multiply, since it has the
earliest exit: if the remainder cannot be perfectly divided by the
operand in question, then we cannot multiply the operands; fallback onto
an addition; additionally, even if the multiplication is applicable, it
might lead to a false path, so we must still try to add the operand.

`remainingOperands` must be converted onto a vector, otherwise the
function will fail to compile, as the view will break the recursion; I
do not fully understand why this is the case.

Deducing this allows to have a recursive lambda that is decorated by
`validateEquation`; this way, the decorator can keep a good API, and
hide the implementation details, i.e. the recursive lambda.