2024 07 Solve part two
Concatenation can only happen if the operand is part of the remainder.
It is not enough if the number can be found in the remainder, but rather
that the component of the operand equal the component of the remainder.
For example, the operand 6, which only has a unit component, is in the
remainder 156, but not in a remainder 165.