2024 08 Create antinodes given two points
While the test input will only work with positive coordinates that make
up the grid, allowing negative numbers abstracts away where it might be
used, while also allowing for an easier check of antinodes that exist in
the grid, that is, antinodes that have negative components do not count.

Both the defaulted comparison and printing of `Point` are declared as
friends to allow compilation, and follow convention respectively. I fail
to see the point in also declaring the remaining functions that work on
`Point` as friends.

Consider that any two antennas make a vector. To create the antinodes
closest to one of those antennas, distance that antenna from the other
by the vector, or its inverse (the first antenna is moved away from the
second by moving it with the inverse vector).

Most expecatations have one antenna at the origin to more easily ensure
the antinodes are in the correct directions. The last expectation does
test that two random locations produce antinodes at the expected places.