2024 08 Parse puzze input
Since the antennas will be used as groups based on their labels, it
makes sense to group them already here, rather than as a separate step.

Unfortunately, I do not know how to push `Point` into `Antennas` while
still piping the input. `std::ranges::to<Antennas>()` does not work, as
it creates a map of `<char, Point>` rather than a vector of `Point`.