2024 08 Find points in grid along a line
Line is defined by two points that are passed as parameters.

The implementation of `findPointsinGridAlongLine()` is not pretty, but
it works, is fast enough, so I fail to see the need to spend more time
on it. It works by iteratively finding the antinode along the direction,
using the last two known antinodes, until the last point is outside the
grid.