2024 08 Find points in grid along a line
Line is defined by two points that are passed as parameters.

Given two points, it is possible to get a function that returns a `y`
for any given `x`; we need only iterate through the entire width, to get
all the valid (x, y) positions. However, a vertical cannot be described
in this way, as the slop is indeterminate (dY/0); for this case we know
that all points share the same `x` component, and need only iterate
throught the height of the grid.